Variable Tracking System is up. Missing a feature, but it's up.
Soooo instead of having a complicated multivariable cross-check
monstrosity, I created an eventTracker which takes in "eventScripts"
which change every NPC in the scene. These eventScripts are passed in
from ConvoNodes. The only feature missing is the TIME component, but
that's trivial to add, it's 4:00 AM, and I'm tired.

Also, I figured out you can add hover ToolTips to variables in the
editor, so I'm DEFINITELY gonna do that from now on.